Modified GPS object to use int32_t instead of doubles because Arduino does not support double floating-point precision.

+++ cpp/testFunctions/status_test.cpp
@@ -8,7 +8,7 @@ int main(){
   printf("%f,%f,%f,%f,%f\n",gps.lat,gps.lon,gps.t,gps.x,gps.y);
   gps.set(45.0,-167.83434234,90.238239);
   printf("%f,%f,%f,%f,%f\n",gps.lat,gps.lon,gps.t,gps.x,gps.y);
-  gps.set(long(-450000000),long(1678343423),90.238239);
+  gps.set(int32_t(-450000000),int32_t(1678343423),90.238239);
   printf("%f,%f,%f,%f,%f\n",gps.lat,gps.lon,gps.t,gps.x,gps.y);
 
   return 0;
